src/models/paddleocr/requirements.txt
# PaddleOCR specific requirements
# Compatible with Python 3.7 (base image)

# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0

# API dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Utilities
requests>=2.25.0
aiofiles>=0.7.0
python-multipart>=0.0.5


